Separa validação de CPF
